--- 04_casting.py
@@ -1,12 +1,0 @@
-#change type of variable types (强制类型转换)
-#the casting does not directly change the type of the varible, instead, it returns a changed type
-x = 1.1
-print(type(x))
-y = int(x)
-print(type(y))
-z = float(y)
-print(type(z))
-w = str(z)
-print(type(w))
-
-
