--- 08_tuple.py
@@ -1,16 +1,0 @@
-#A tuple is a collection that is ordered and unchangeable
-
-#Constructor tuple()
-tuple_new = tuple(("apple", "banana", "cherry"))
-print(tuple_new)
-
-#Accessing method is the same as List. The difference is that members are unchangeable.
-#add, remove do not work with tuples
-print(tuple_new[1])
-
-
-#len and other methods that do not change the tupke still work
-
-#count()	Returns the number of times a specified value occurs in a tuple
-#index()	Searches the tuple for a specified value and returns the position of where it was found
-pass
