--- 11_conditional.py
@@ -1,17 +1,0 @@
-#Conditional syntax is similar to that in C++
-#except that else if is replaced by elif
-#break and continue work in the same way
-#case is not available in python
-a = 1
-b = 2
-if a > b:
-    print("a > b")
-elif a == b:
-    print("a = b")
-else:
-    print("a < b")
-
-
-
-
-
