--- datafile_1
@@ -1,15 +1,0 @@
-DPD
-VER
-DOC
-TST
-BRC
-NFT
-MOD
-DEL
-FIX
-ANT
-RFT
-IMP
-RMN
-CLN
-CMT
