--- datafile_2
@@ -1,15 +1,0 @@
-18
-14
-15
-5
-3
-121
-58
-22
-54
-5
-28
-1
-16
-10
-2
