--- datafile_3
@@ -1,15 +1,0 @@
-6
-13
-45
-7
-1
-174
-118
-23
-157
-16
-22
-8
-34
-26
-18
