+++ leetcode/_001_twosum.py
@@ -1,0 +1,16 @@
+from typing import List
+import numpy as np
+
+"""
+Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
+
+You may assume that each input would have exactly one solution, and you may not use the same element twice.
+
+You can return the answer in any order.
+"""
+
+class Solution:
+    def twoSum(self, nums: List[int], target: int) -> List[int]:
+        
+        return
+
