+++ leetcode/solution.py
@@ -1,0 +1,10 @@
+from _001_twosum import Solution
+
+sol = Solution()
+list_in = [2,7,11,15]
+print(list_in)
+list_out = sol.twoSum(list_in, 9)
+
+print(list_out)
+
+
