+++ tutorial/01_IO_syntax.py
@@ -1,0 +1,20 @@
+#python use indentation to indicate a code block
+class my_class(object):
+    print("Hello World!")
+    #pass is used to keep integrity of code, doing nothing. 
+    #Its usage is like "continue" in C++ but it could also be applied to classes and functions.
+    pass
+
+#print is used to output variables where '+' could be used either as a combinator of different types 
+#or a operator to add two numbers
+    print(x, y, z, sep = ', ')
+
+#input()
+#input from command-line
+x = input()
+print(x)
+
+#python use # to add comment to code, """ """ is a Docstring which could also be used as comment.
+"""This a Doc String"""
+
+
