+++ tutorial/02_variable.py
@@ -1,0 +1,25 @@
+#Python does not have command for declaring a variable
+#Variables are created the moment when they are assigned values
+
+#str
+carname = "Volvo" 
+print(carname)
+
+#int
+x = 5
+y = 10
+z = x + y
+print(x)
+
+#boolean
+choice = False
+
+#float
+pi = 3.14
+
+#naming rules
+#A variable name must start with a letter or the underscore character
+#A variable name cannot start with a number
+#A variable name can only contain alpha-numeric characters and underscores (A-z, 0-9, and _ )
+#Variable names are case-sensitive (age, Age and AGE are three different variables)
+
