+++ tutorial/03_numbers.py
@@ -1,0 +1,8 @@
+x = 1   #int
+y = 1.1 #float
+z = 1j  #complex
+
+#use type() to verify the variable type
+print(type(x))
+print(type(y))
+print(type(z))
