+++ tutorial/06_operator.py
@@ -1,0 +1,37 @@
+#Arithmetic(Assignment) Operators
+#=
+#addition       +(=)
+#subtraction    -(=)
+#multiplication *(=)
+#division       /(=)
+#modulus        %(=)
+#Exponentiation **(=)
+#Floor division //(=)
+
+#Bitwise(Assignment) Operators
+#&(=)
+#|(=)
+#^(=)
+#>>(=)
+#<<(=) 
+
+#Comparison Operators
+#==
+#!=
+#>
+#<
+#>=
+#<= 
+
+#Logical Operators
+#and
+#or
+#not() 
+
+#Identity Operator, comparing objects, both in value and in physical memory location
+#is
+#is not 
+
+#Membership Operator, whether a sequence is in a object or not
+#in
+#not in
