--- utils.py
@@ -1,8 +1,0 @@
-class Student:
-    def __init__(self, name, age):
-        self.name = name
-        self.age = age
-        print("I am " + name + " and I am " + age)
-
-def add(a, b):
-    return (a + b)
